# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a RESTful API for Effective-Mobile project",
        "title": "Effective-Mobile API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/cars": {
            "post": {
                "description": "Add a new car to the database from JSON input.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Functions"
                ],
                "summary": "Create car",
                "operationId": "create-car",
                "parameters": [
                    {
                        "description": "Car info (only regnum, mark, model, year and owner_id are required).",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.Car"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusOKMessage"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusBadRequestMessage"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusNotFoundMessage"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusInternalServerErrorMessage"
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "put": {
                "description": "Update car info in the database by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Functions"
                ],
                "summary": "Update car",
                "operationId": "update-car",
                "parameters": [
                    {
                        "description": "Car info to update (only regnum, mark, model, year and owner_id are required).",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.Car"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusOKMessage"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusBadRequestMessage"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusNotFoundMessage"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusInternalServerErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete car from database by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Functions"
                ],
                "summary": "Delete car",
                "operationId": "delete-car",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusOKMessage"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusBadRequestMessage"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusNotFoundMessage"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusInternalServerErrorMessage"
                        }
                    }
                }
            }
        },
        "/info": {
            "get": {
                "description": "Get all cars from database or filtered by regnum, mark and model (optional).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Functions"
                ],
                "summary": "Get cars",
                "operationId": "get-all-cars",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusOKMessage"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusBadRequestMessage"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusNotFoundMessage"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusInternalServerErrorMessage"
                        }
                    }
                }
            }
        },
        "/owners": {
            "post": {
                "description": "Add a new owner to the database from JSON input.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Functions"
                ],
                "summary": "Add owner",
                "operationId": "add-owner",
                "parameters": [
                    {
                        "description": "Owner info (only name, surname are required).",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.Owner"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusOKMessage"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusBadRequestMessage"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusNotFoundMessage"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/Effective-Mobile_internal_structures.StatusInternalServerErrorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Effective-Mobile_internal_structures.Car": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "mark": {
                    "type": "string",
                    "example": "BMW"
                },
                "model": {
                    "type": "string",
                    "example": "X5"
                },
                "owner_id": {
                    "type": "integer",
                    "example": 1
                },
                "regnum": {
                    "type": "string",
                    "example": "A111AA155"
                },
                "year": {
                    "type": "integer",
                    "example": 2015
                }
            }
        },
        "Effective-Mobile_internal_structures.Owner": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Max"
                },
                "patronymic": {
                    "type": "string",
                    "example": "Olegovich"
                },
                "surname": {
                    "type": "string",
                    "example": "Shestakov"
                }
            }
        },
        "Effective-Mobile_internal_structures.StatusBadRequestMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error: bad request"
                }
            }
        },
        "Effective-Mobile_internal_structures.StatusInternalServerErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error: internal server error"
                }
            }
        },
        "Effective-Mobile_internal_structures.StatusNotFoundMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error: not found"
                }
            }
        },
        "Effective-Mobile_internal_structures.StatusOKMessage": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    }
}